Move `AffineMatrixBackground` to the `affine_background` module
